# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SubPaperFlux API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/site-configs": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Create a site config",
        "operationId": "create_site_config_v1_v1_site_configs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/SiteConfigSelenium"
                  },
                  {
                    "$ref": "#/components/schemas/SiteConfigApi"
                  }
                ],
                "discriminator": {
                  "propertyName": "login_type",
                  "mapping": {
                    "selenium": "#/components/schemas/SiteConfigSelenium",
                    "api": "#/components/schemas/SiteConfigApi"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SiteConfigSeleniumOut"
                    },
                    {
                      "$ref": "#/components/schemas/SiteConfigApiOut"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "login_type",
                    "mapping": {
                      "selenium": "#/components/schemas/SiteConfigSeleniumOut",
                      "api": "#/components/schemas/SiteConfigApiOut"
                    }
                  },
                  "title": "Response Create Site Config V1 V1 Site Configs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "List site configs",
        "operationId": "list_site_configs_v1_v1_site_configs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_global",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Global"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteConfigsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/site-configs/{config_id}": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Retrieve a site config",
        "operationId": "get_site_config_v1_v1_site_configs__config_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SiteConfigSeleniumOut"
                    },
                    {
                      "$ref": "#/components/schemas/SiteConfigApiOut"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "login_type",
                    "mapping": {
                      "selenium": "#/components/schemas/SiteConfigSeleniumOut",
                      "api": "#/components/schemas/SiteConfigApiOut"
                    }
                  },
                  "title": "Response Get Site Config V1 V1 Site Configs  Config Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "v1"
        ],
        "summary": "Update a site config",
        "operationId": "update_site_config_v1_v1_site_configs__config_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/SiteConfigSelenium"
                  },
                  {
                    "$ref": "#/components/schemas/SiteConfigApi"
                  }
                ],
                "discriminator": {
                  "propertyName": "login_type",
                  "mapping": {
                    "selenium": "#/components/schemas/SiteConfigSelenium",
                    "api": "#/components/schemas/SiteConfigApi"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SiteConfigSeleniumOut"
                    },
                    {
                      "$ref": "#/components/schemas/SiteConfigApiOut"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "login_type",
                    "mapping": {
                      "selenium": "#/components/schemas/SiteConfigSeleniumOut",
                      "api": "#/components/schemas/SiteConfigApiOut"
                    }
                  },
                  "title": "Response Update Site Config V1 V1 Site Configs  Config Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "v1"
        ],
        "summary": "Delete a site config",
        "operationId": "delete_site_config_v1_v1_site_configs__config_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/site-configs/{config_id}/test": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Test site config selectors against the login page",
        "operationId": "test_site_config_v1_site_configs__config_id__test_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Test Site Config V1 Site Configs  Config Id  Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/site-configs/{config_id}/copy": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Copy Site Config V1",
        "operationId": "copy_site_config_v1_v1_site_configs__config_id__copy_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SiteConfigSeleniumOut"
                    },
                    {
                      "$ref": "#/components/schemas/SiteConfigApiOut"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "login_type",
                    "mapping": {
                      "selenium": "#/components/schemas/SiteConfigSeleniumOut",
                      "api": "#/components/schemas/SiteConfigApiOut"
                    }
                  },
                  "title": "Response Copy Site Config V1 V1 Site Configs  Config Id  Copy Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SiteConfigSelenium": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "site_url": {
            "type": "string",
            "minLength": 1,
            "format": "uri",
            "title": "Site Url"
          },
          "owner_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner User Id"
          },
          "success_text_class": {
            "type": "string",
            "title": "Success Text Class",
            "default": ""
          },
          "expected_success_text": {
            "type": "string",
            "title": "Expected Success Text",
            "default": ""
          },
          "required_cookies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Cookies"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "login_type": {
            "type": "string",
            "const": "selenium",
            "title": "Login Type",
            "default": "selenium"
          },
          "selenium_config": {
            "$ref": "#/components/schemas/SeleniumConfig"
          }
        },
        "type": "object",
        "required": [
          "name",
          "site_url",
          "selenium_config"
        ],
        "title": "SiteConfigSelenium"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}